asr/requirements.txt
anyio
fastapi
requests
google-generativeai
pydantic_settings
